{"schema":"decodex/commit/1","summary":"Extract toolbar pointer motion owner","authority":"manual"}
